Update api/user endpoint with feedback from Jason
